add display bill feature

+++ database/migrations/2019_03_24_104411_create_bill_transactions_table.php
@@ -21,7 +21,7 @@ class CreateBillTransactionsTable extends Migration
             $table->foreign('bill_participant_id_from')->references('id')->on('bill_participants')->onDelete('cascade');
             $table->integer('bill_participant_id_to')->unsigned();
             $table->foreign('bill_participant_id_to')->references('id')->on('bill_participants')->onDelete('cascade');
-            $table->integer('amount')->unsigned();
+            $table->float('amount')->unsigned();
             $table->timestamps();
         });
